backend/requirements.txt
et-xmlfile==1.1.0
iniconfig==2.0.0
numpy==1.26.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
pluggy==1.3.0
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
tzdata==2023.3
